third_party/blink/web_tests/platform/mac/external/wpt/css/css-text/text-align/text-align-last-empty-inline-expected.txt

This is a testharness.js-based test.
FAIL content that cannot be justified should be centered when text-align-last is justify assert_equals: expected 395.4375 but got 8
Harness: the test ran to completion.
